Changed handling of newtonsoft assembly
