+++ Commands/ClientSidePages/SaveClientSidePageConversionLog.cs
@@ -18,12 +18,12 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
     {
         private Assembly modernizationAssembly;
         private Assembly sitesCoreAssembly;
-        private Assembly newtonsoftAssembly;
+        //private Assembly newtonsoftAssembly;
 
         protected override void ExecuteCmdlet()
         {
             //Fix loading of modernization framework
-            FixAssemblyResolving();
+            //FixAssemblyResolving();
 
             // Get last used transformator instance from cache
             var transformator = CacheManager.Instance.GetLastUsedTransformator();
@@ -34,45 +34,45 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
             }
         }
 
-        private string AssemblyDirectory
-        {
-            get
-            {
-                string codeBase = Assembly.GetExecutingAssembly().CodeBase;
-                UriBuilder uri = new UriBuilder(codeBase);
-                string path = Uri.UnescapeDataString(uri.Path);
-                return Path.GetDirectoryName(path);
-            }
-        }
+        //private string AssemblyDirectory
+        //{
+        //    get
+        //    {
+        //        string codeBase = Assembly.GetExecutingAssembly().CodeBase;
+        //        UriBuilder uri = new UriBuilder(codeBase);
+        //        string path = Uri.UnescapeDataString(uri.Path);
+        //        return Path.GetDirectoryName(path);
+        //    }
+        //}
 
-        private void FixAssemblyResolving()
-        {
-            try
-            {
-                newtonsoftAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "NewtonSoft.Json.dll"));
-                sitesCoreAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "OfficeDevPnP.Core.dll"));
-                modernizationAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "SharePointPnP.Modernization.Framework.dll"));
-                AppDomain.CurrentDomain.AssemblyResolve += CurrentDomain_AssemblyResolve;
-            }
-            catch { }
-        }
+        //private void FixAssemblyResolving()
+        //{
+        //    try
+        //    {
+        //        newtonsoftAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "NewtonSoft.Json.dll"));
+        //        sitesCoreAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "OfficeDevPnP.Core.dll"));
+        //        modernizationAssembly = Assembly.LoadFrom(Path.Combine(AssemblyDirectory, "SharePointPnP.Modernization.Framework.dll"));
+        //        AppDomain.CurrentDomain.AssemblyResolve += CurrentDomain_AssemblyResolve;
+        //    }
+        //    catch { }
+        //}
 
-        private Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
-        {
-            if (args.Name.StartsWith("OfficeDevPnP.Core"))
-            {
-                return sitesCoreAssembly;
-            }
-            if (args.Name.StartsWith("Newtonsoft.Json"))
-            {
-                return newtonsoftAssembly;
-            }
-            if (args.Name.StartsWith("SharePointPnP.Modernization.Framework"))
-            {
-                return modernizationAssembly;
-            }
-            return null;
-        }
+        //private Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
+        //{
+        //    if (args.Name.StartsWith("OfficeDevPnP.Core"))
+        //    {
+        //        return sitesCoreAssembly;
+        //    }
+        //    if (args.Name.StartsWith("Newtonsoft.Json"))
+        //    {
+        //        return newtonsoftAssembly;
+        //    }
+        //    if (args.Name.StartsWith("SharePointPnP.Modernization.Framework"))
+        //    {
+        //        return modernizationAssembly;
+        //    }
+        //    return null;
+        //}
 
     }
 }
